[#816] Consolidate domain management into store for better cohesion
• Removes useDomains composable
• Enhances domainsStore with toggle feature
• Simplifies domainsTable implementation
• Ensures brand settings are preserved
• Improves error handling flow
