+++ src/stores/domainsStore.ts
@@ -35,6 +35,40 @@ export const useDomainsStore = defineStore('domains', {
       }
     },
 
+
+
+    async toggleHomepageAccess(domain: CustomDomain) {
+      const newHomepageStatus = !domain?.brand?.allow_public_homepage;
+
+      try {
+        await api.put(`/api/v2/account/domains/${domain.display_domain}/brand`, {
+          brand: { allow_public_homepage: newHomepageStatus }
+        });
+
+        // Ensure we maintain all required fields from the original brand settings
+        if (domain.brand) {
+          this.updateDomain({
+            ...domain,
+            brand: {
+              primary_color: domain.brand.primary_color,
+              instructions_pre_reveal: domain.brand.instructions_pre_reveal,
+              instructions_reveal: domain.brand.instructions_reveal,
+              instructions_post_reveal: domain.brand.instructions_post_reveal,
+              button_text_light: domain.brand.button_text_light,
+              font_family: domain.brand.font_family,
+              corner_style: domain.brand.corner_style,
+              allow_public_homepage: newHomepageStatus
+            }
+          });
+        }
+
+        return newHomepageStatus;
+      } catch (error) {
+        console.error('Failed to toggle homepage access:', error);
+        throw error;
+      }
+    },
+
     addDomain(domain: CustomDomain) {
       this.domains.push(domain);
     },
